Updated modules/utils.js (added 1 line, deleted 1 line).

+++ modules/utils.js
@@ -7,7 +7,7 @@ Alec L. Robitaille
 
 // Set year
 var set_year = function(img) {
-	return img.set('year', img.date().get('year'));
+	return img.set('year', img.date().get('year').int());
 };
 exports.set_year = set_year;
